fps tests

+++ src/dataset/RTTFibers.h
@@ -79,7 +79,7 @@ public:
 
 
 	unsigned int  m_trackActionStep;
-	unsigned int m_timerStep;
+	float m_timerStep;
